requirements.txt
attrs~=17.4.0
numpy~=1.22.3
enum34~=1.1.6
matplotlib~=3.1.2
pynput~=1.7.3
tf~=1.13.2
rospy~=1.15.14
requests~=2.22.0
click~=7.0
websocket-client~=0.59.0
PyYAML~=5.3.1
setuptools~=62.1.0
roslib~=1.15.8
rostopic~=1.15.14
mujoco~=2.1.5
scipy~=1.8.0
glfw~=2.5.3
imageio~=2.16.1